requirements.txt
discord.py
httpx
fastapi[all]
asyncpg
